Refactor vote-type handling
